Finished primary consistency checks with POPART-IBM repository.
Codes are consistent (except when they are meant to differ -
 - for POPART-IBM: CEA, TREATS;
 - for Manicaland: PrEP, HSV-2, MTCT, separating out PopART and non-PopART things more clearly (flag IS_POPART, moving PopART parameters to a separate function in input.c, adding cohort snapshot)

To do: fix the output function store_timesteps_output(). I've modified this function to use the age_list (rather than going through the whole population).
The advantage is that it speeds up the code. However, it does not properly output the PC population, because of how we define it in the code.
So I need to add a flag/separate bit of code to deal specifically with the PC population.