Added tool check_HPC_files.py that can compare the .c and .h files to a template directory's files. It also checks that the constants in constants.h are set to zero so as to not generate unwanted output files.
HPC_5Oct2021_R1_R6.Rnw is a sweave file containing the plots for the impact (calibration plots, plus validation and impact plots of the calibrated files).

Fixed bug where condom use was modified in 2021 regardless of whether there was an intervention (basically we called the function update_partnership_condom_use_in_response_to_intervention() even when there wasn't an intervention. This is now checked in simul.c so the function si not called).